adds PaymentHandler
